utils: Refactor Python script into multiple scripts
Refactor `all.py` script into multiple scripts. Each script stores its
own class: `TesterConfig`, `SystemConfig`, `AppConfig`, `TargetSetup`,
`BuildSetup`, `RunSetup`.
